Refactor assessments API for clarity
