+++ backend/courses/assessments/api/__init__.py
@@ -1,0 +1,6 @@
+from courses.assessments.api.get_assessment import get_assessment
+from courses.assessments.api.complete_assessment import complete_assessment
+from courses.assessments.api.submit_checkbox_answer import submit_checkbox_answer
+from courses.assessments.api.submit_coding_answer import submit_coding_answer
+from courses.assessments.api.submit_multiple_choice_answer import submit_multiple_choice_answer
+from courses.assessments.api.submit_written_response_answer import submit_written_response_answer
