+++ backend/tests/courses/assessments/test_submit_checkbox_answer.py
@@ -5,10 +5,7 @@ from tests.utils import (
 )
 from django.contrib.auth.models import User
 import courses.models as db
-from datetime import (
-    datetime,
-    timezone
-)
+from django.utils import timezone
 from rest_framework import status
 from uuid import (
     UUID,
@@ -18,7 +15,7 @@ from uuid import (
 
 class AnswerCheckboxQuestion(TestCasesWithUserAuth):
     def get_api_endpoint(self, assessment_id: UUID, checkbox_question_id: UUID) -> str:
-        return f'/api/v0/assessments/{assessment_id}/answer_question/checkbox/{str(checkbox_question_id)}/'
+        return f'/api/v0/assessments/{assessment_id}/answer/checkbox/{str(checkbox_question_id)}/'
     
     def test_no_existing_answer_obj_happy_path(self):
         requesting_user_id = self.user.id
@@ -308,3 +305,51 @@ class AnswerCheckboxQuestion(TestCasesWithUserAuth):
         
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
         self.assertDictEqual(response.json(), expected_body)
+
+    def test_submission_after_assessment_completed_throws_error(self):
+        requesting_user_id = self.user.id
+        
+        assessment = create_assessment(user_id=requesting_user_id)
+        
+        assessment_submission = create_assessment_submission(
+            user_id=requesting_user_id,
+            assessment_id=assessment.id
+        )
+        
+        checkbox_question = db.CheckboxQuestion.objects.create(
+            assessment=assessment,
+            prompt='Choose all positive numbers',
+            order=2,
+            points=4,
+            options=['-1', '-2', '3', '4'],
+            correct_option_indices=[2, 3]
+        )
+        
+        checkbox_answer = db.CheckboxAnswer.objects.create(
+            assessment_submission=assessment_submission,
+            question=checkbox_question,
+            selected_answer_indices=[0, 1, 2]
+        )
+
+        assessment_submission.completed_at = timezone.now()
+        assessment_submission.save()
+        
+        data = {
+            'selected_answer_indices': [2, 3]
+        }
+
+        response = self.client.post(
+            self.get_api_endpoint(
+                assessment_id=assessment.id,
+                checkbox_question_id=checkbox_question.id
+            ), data=data
+        )
+
+        expected_body = {'error': 'The assessment has already been completed'}
+        
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertDictEqual(response.json(), expected_body)
+        
+        checkbox_answer.refresh_from_db()
+        
+        self.assertEqual(checkbox_answer.selected_answer_indices, [0, 1, 2])
