+++ backend/tests/courses/assessments/test_submit_multiple_choice_answer.py
@@ -5,10 +5,7 @@ from tests.utils import (
 )
 from django.contrib.auth.models import User
 import courses.models as db
-from datetime import (
-    datetime,
-    timezone
-)
+from django.utils import timezone
 from rest_framework import status
 from uuid import (
     UUID,
@@ -18,7 +15,7 @@ from uuid import (
 
 class AnswerMultipleChoiceQuestion(TestCasesWithUserAuth):
     def get_api_endpoint(self, assessment_id: UUID, multiple_choice_question_id: UUID) -> str:
-        return f'/api/v0/assessments/{assessment_id}/answer_question/multiple_choice/{str(multiple_choice_question_id)}/'
+        return f'/api/v0/assessments/{assessment_id}/answer/multiple_choice/{str(multiple_choice_question_id)}/'
     
     def test_no_existing_answer_obj_happy_path(self):
         requesting_user_id = self.user.id
@@ -251,3 +248,56 @@ class AnswerMultipleChoiceQuestion(TestCasesWithUserAuth):
         
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
         self.assertDictEqual(response.json(), expected_body)
+
+    def test_submission_after_assessment_completed_throws_error(self):
+        requesting_user_id = self.user.id
+        
+        assessment = create_assessment(user_id=requesting_user_id)
+        
+        assessment_submission = create_assessment_submission(
+            user_id=requesting_user_id,
+            assessment_id=assessment.id
+        )
+        
+        multiple_choice_question = db.MultipleChoiceQuestion.objects.create(
+            assessment=assessment,
+            prompt='Choose the animal that flies',
+            order=1,
+            points=3,
+            options=[
+                'Deer',
+                'Bird',
+                'Shark'
+            ],
+            correct_option_index=1
+        )
+        
+        multiple_choice_answer = db.MultipleChoiceAnswer.objects.create(
+            assessment_submission=assessment_submission,
+            question=multiple_choice_question,
+            selected_answer_index=0
+        )
+        
+        assessment_submission.completed_at = timezone.now()
+        assessment_submission.save()
+
+        data = {
+            'selected_answer_index': 1
+        }
+
+        response = self.client.post(
+            self.get_api_endpoint(
+                assessment_id=assessment.id,
+                multiple_choice_question_id=multiple_choice_question.id
+            ), data=data
+        )
+        
+        expected_body = {'error': 'The assessment has already been completed'}
+        
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertDictEqual(response.json(), expected_body)
+        
+        multiple_choice_answer.refresh_from_db()
+        
+        self.assertEqual(multiple_choice_answer.selected_answer_index, 0)
+
